+++ chain/services/miner_service.go
@@ -9,8 +9,8 @@ package services
 
 import (
 	"errors"
+	"github.com/qlcchain/go-qlc/consensus/pov"
 
-	"github.com/qlcchain/go-qlc/consensus"
 	"github.com/qlcchain/go-qlc/log"
 	"github.com/qlcchain/go-qlc/miner"
 	"go.uber.org/zap"
@@ -25,7 +25,7 @@ type MinerService struct {
 	logger *zap.SugaredLogger
 }
 
-func NewMinerService(cfg *config.Config, povEngine *consensus.PoVEngine) *MinerService {
+func NewMinerService(cfg *config.Config, povEngine *pov.PoVEngine) *MinerService {
 	miner := miner.NewMiner(cfg, povEngine)
 	return &MinerService{
 		miner:  miner,
