+++ consensus/pov/pov_chain.go
@@ -1,4 +1,4 @@
-package consensus
+package pov
 
 import (
 	"errors"
