+++ consensus/pov/pov_processor.go
@@ -1,4 +1,4 @@
-package consensus
+package pov
 
 import (
 	"sync"
