+++ consensus/pov/pov_state.go
@@ -1,4 +1,4 @@
-package consensus
+package pov
 
 import (
 	"errors"
