+++ consensus/pov/pov_sync.go
@@ -1,4 +1,4 @@
-package consensus
+package pov
 
 import (
 	"math/big"
